requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow
opencv-python-headless
numpy
Pillow
